backend/requirements.txt
# LangChain backend for portable AI Agent
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
openai>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
email-validator>=2.0.0
